+++ frontend/src/components/Common/ImageUploader/ImageUploader.stories.tsx
@@ -1,0 +1,13 @@
+import type { Meta, StoryObj } from '@storybook/react';
+
+import ImageUploader from './ImageUploader';
+
+const meta: Meta<typeof ImageUploader> = {
+  title: 'common/ImageUploader',
+  component: ImageUploader,
+};
+
+export default meta;
+type Story = StoryObj<typeof ImageUploader>;
+
+export const Default: Story = {};
