--- frontend/src/components/Review/ReviewImageUploader/ReviewImageUploader.stories.tsx
@@ -1,13 +1,0 @@
-import type { Meta, StoryObj } from '@storybook/react';
-
-import ReviewImageUploader from './ReviewImageUploader';
-
-const meta: Meta<typeof ReviewImageUploader> = {
-  title: 'review/ReviewImageUploader',
-  component: ReviewImageUploader,
-};
-
-export default meta;
-type Story = StoryObj<typeof ReviewImageUploader>;
-
-export const Default: Story = {};
